Adding Helm chart
Adding a Helm chart to facilitate K8s deployment of the NLP evaluation
service. The pom has also been supplemented with a helm-maven-plugin to
facilitate packaging of the Helm chart for distribution to a chart
repository. The chart permits user specification of a static IP address
for the service. Implementation of the K8s service as a load balancer
requires the static IP to be ‘regional’ (at least on GCP). Resolves #3.